+++ views/background.js
@@ -6,7 +6,7 @@ const BIBLE_TABS_URL = [
   "https://www.bible.com/bible*",
   "https://www.bible.com/*/bible*"
 ];
-const DEFAULT_URL = "https://my.bible.com/bible";
+const DEFAULT_URL = "https://www.bible.com/bible";
 
 const projectorPage = "views/projector/tab.html";
 const settingsPage = "views/settings/options.html";